Create mmn 11
Write a plan for calculating the circumference and area of ​​a triangle.
The plot will read from the input that the pocket uses the length of the three sides of the triangle, and print 
Screen the perimeter of the triangle and its area. Rib lengths will be natural (i.e. integer) numbers
Positive.
Note that the user may enter three numbers that cannot be the side lengths of a triangle
(For example: 1, 2, 4 - why?). 
You need to check this and let us know if it did happen.
To calculate the area of ​​the triangle, you can use a Heron deviation that states that the area of ​​the triangle is equal to the root.
The square of (c-s) (b-s) (a-s (s, where a, b and c are the lengths of the three sides of the triangle, and s is half scope.
You can find Heron's distraction on Wikipedia at:
http://en.wikipedia.org/wiki/%D7%A0%D7%95%D7%A1%D7%97%D7%AA_%D7%94%D
 7% A8% D7% 95% D7% 9F
You should check the input frequencies. If the input is incorrect, the program should print the input characters accompanied by appropriate messages.
The plot you wrote should be in a department called Triangle.
For your help, write parts of the department here. You must fill in the blanks (including the missing documentation).